feat: refine user center feature
