+++ console/package.json
@@ -119,6 +119,7 @@
     "@types/lodash.debounce": "^4.0.7",
     "@types/lodash.isequal": "^4.5.6",
     "@types/lodash.merge": "^4.6.7",
+    "@types/lodash.sortby": "^4.7.9",
     "@types/node": "^18.11.19",
     "@types/qs": "^6.9.7",
     "@types/randomstring": "^1.1.8",
