+++ console/uc-src/main.ts
@@ -3,7 +3,7 @@ import { createApp, type DirectiveBinding } from "vue";
 import App from "./App.vue";
 import { setupVueQuery } from "@/setup/setupVueQuery";
 import { setupComponents } from "@/setup/setupComponents";
-import { setupI18n } from "@/locales";
+import { getBrowserLanguage, i18n, setupI18n } from "@/locales";
 import { createPinia } from "pinia";
 import { setupCoreModules, setupPluginModules } from "@uc/setup/setupModules";
 import router from "@uc/router";
@@ -11,6 +11,7 @@ import { useUserStore } from "@/stores/user";
 import { apiClient } from "@/utils/api-client";
 import { useRoleStore } from "@/stores/role";
 import { hasPermission } from "@/utils/permission";
+import { useGlobalInfoStore } from "@/stores/global-info";
 
 const app = createApp(App);
 
@@ -55,19 +56,34 @@ async function loadUserPermissions() {
 })();
 
 async function initApp() {
-  setupCoreModules(app);
-
-  const userStore = useUserStore();
-  await userStore.fetchCurrentUser();
-
-  loadUserPermissions();
-
   try {
-    await setupPluginModules(app);
-  } catch (e) {
-    console.error("Failed to load plugins", e);
-  }
+    setupCoreModules(app);
 
-  app.use(router);
-  app.mount("#app");
+    const userStore = useUserStore();
+    await userStore.fetchCurrentUser();
+
+    // set locale
+    i18n.global.locale.value =
+      localStorage.getItem("locale") || getBrowserLanguage();
+
+    const globalInfoStore = useGlobalInfoStore();
+    await globalInfoStore.fetchGlobalInfo();
+
+    if (userStore.isAnonymous) {
+      return;
+    }
+
+    await loadUserPermissions();
+
+    try {
+      await setupPluginModules(app);
+    } catch (e) {
+      console.error("Failed to load plugins", e);
+    }
+  } catch (error) {
+    console.error("Failed to init app", error);
+  } finally {
+    app.use(router);
+    app.mount("#app");
+  }
 }
