+++ console/uc-src/modules/profile/module.ts
@@ -1,6 +1,6 @@
 import { definePlugin } from "@halo-dev/console-shared";
 import BasicLayout from "@uc/layouts/BasicLayout.vue";
-import { IconUserLine } from "@halo-dev/components";
+import { IconAccountCircleLine } from "@halo-dev/components";
 import { markRaw } from "vue";
 import Profile from "./Profile.vue";
 
@@ -21,7 +21,7 @@ export default definePlugin({
             searchable: true,
             menu: {
               name: "我的",
-              icon: markRaw(IconUserLine),
+              icon: markRaw(IconAccountCircleLine),
               priority: 0,
               mobile: true,
             },
