+++ console/uc-src/router/guards/auth-check.ts
@@ -1,0 +1,17 @@
+import { useUserStore } from "@/stores/user";
+import type { Router } from "vue-router";
+
+export function setupAuthCheckGuard(router: Router) {
+  router.beforeEach((to, from, next) => {
+    const userStore = useUserStore();
+
+    if (userStore.isAnonymous) {
+      window.location.href = `/console/login?redirect_uri=${encodeURIComponent(
+        window.location.href
+      )}`;
+      return;
+    }
+
+    next();
+  });
+}
